chore: move structure and field config to service folder
