+++ src/Facades/FieldConfiguration.php
@@ -5,14 +5,19 @@ declare(strict_types=1);
 namespace Innobrain\Structure\Facades;
 
 use Illuminate\Support\Facades\Facade;
+use Innobrain\OnOfficeAdapter\Dtos\OnOfficeApiCredentials;
+use Innobrain\Structure\Collections\ModulesCollection;
+use Innobrain\Structure\Services\FieldConfiguration as ServiceFieldConfiguration;
 
 /**
- * @see \Innobrain\Structure\FieldConfiguration
+ * @see ServiceFieldConfiguration
+ *
+ * @method static ModulesCollection retrieveForClient(OnOfficeApiCredentials $credentials, array $only = [])
  */
 class FieldConfiguration extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
-        return \Innobrain\Structure\FieldConfiguration::class;
+        return ServiceFieldConfiguration::class;
     }
 }
